Desktop/Coding Projects/affordability/requirements.txt
fredapi
pandas
plotly
python-dotenv
